django_part/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==4.1.2
django-crispy-forms==1.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
MarkupSafe==2.1.1
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pytz==2022.4
sqlparse==0.4.3
Werkzeug==2.2.2
